+++ packages/headless/doc-parser/src/interface-resolver.test.ts
@@ -1,30 +1,19 @@
-import {
-  ApiMethodSignature,
-  ExcerptTokenKind,
-  IExcerptToken,
-  ReleaseTag,
-} from '@microsoft/api-extractor-model';
+import {ApiMethodSignature, ReleaseTag} from '@microsoft/api-extractor-model';
+import {buildMockApiDocComment} from '../mocks/mock-api-doc-comment';
 import {buildMockApiInterface} from '../mocks/mock-api-interface';
 import {buildMockApiPropertySignature} from '../mocks/mock-api-property-signature';
 import {buildMockApiTypeAlias} from '../mocks/mock-api-type-alias';
 import {buildMockEntity} from '../mocks/mock-entity';
 import {buildMockEntryPoint} from '../mocks/mock-entry-point';
+import {
+  buildContentExcerptToken,
+  buildReferenceExcerptToken,
+} from '../mocks/mock-excerpt-token';
 import {buildMockFuncEntity} from '../mocks/mock-func-entity';
 import {buildMockObjEntity} from '../mocks/mock-obj-entity';
 import {resolveInterfaceMembers} from './interface-resolver';
 
 describe('#resolveInterfaceMembers', () => {
-  function buildContentExcerptToken(text: string): IExcerptToken {
-    return {kind: ExcerptTokenKind.Content, text};
-  }
-
-  function buildReferenceExcerptToken(
-    text: string,
-    canonicalReference: string
-  ): IExcerptToken {
-    return {kind: ExcerptTokenKind.Reference, text, canonicalReference};
-  }
-
   it('resolves a property with a primitive type', () => {
     const entryPoint = buildMockEntryPoint();
     const apiInterface = buildMockApiInterface({name: 'Pager'});
@@ -103,8 +92,12 @@ describe('#resolveInterfaceMembers', () => {
     const entryPoint = buildMockEntryPoint();
     const pagerInterface = buildMockApiInterface({name: 'Pager'});
 
+    const docComment = buildMockApiDocComment(
+      '/**\n * Returns `true` when the current page is equal to the passed page, and `false` otherwise.\n *\n * @param page - The page number to check.\n *\n * @returns Whether the passed page is selected.\n */\n'
+    );
+
     const method = new ApiMethodSignature({
-      docComment: undefined,
+      docComment,
       excerptTokens: [
         buildContentExcerptToken('isCurrentPage(page: '),
         buildContentExcerptToken('number'),
@@ -130,9 +123,18 @@ describe('#resolveInterfaceMembers', () => {
     entryPoint.addMember(pagerInterface);
 
     const result = resolveInterfaceMembers(entryPoint, pagerInterface);
+
+    const paramEntity = buildMockEntity({
+      name: 'page',
+      type: 'number',
+      desc: 'The page number to check.',
+    });
+
     const funcEntity = buildMockFuncEntity({
       name: 'isCurrentPage',
-      params: [buildMockEntity({name: 'page', type: 'number'})],
+      desc:
+        'Returns `true` when the current page is equal to the passed page, and `false` otherwise.',
+      params: [paramEntity],
       returnType: 'boolean',
     });
     expect(result).toEqual([funcEntity]);