+++ packages/headless/src/controllers/index.ts
@@ -50,11 +50,13 @@ export {
 
 export {
   buildNumericRange,
+  NumericRangeOptions,
+  NumericRangeRequest,
+  buildNumericFacet,
   NumericFacetOptions,
   NumericFacetProps,
   NumericFacetState,
   NumericFacet,
-  buildNumericFacet,
 } from './facets/range-facet/numeric-facet/headless-numeric-facet';
 
 export {History, HistoryState, buildHistory} from './history/headless-history';
@@ -85,6 +87,13 @@ export {
   ResultList,
   buildResultList,
 } from './result-list/headless-result-list';
+
+export {
+  InteractiveResultOptions,
+  InteractiveResultProps,
+  InteractiveResult,
+  buildInteractiveResult,
+} from './result-list/headless-interactive-result';
 
 export {
   ResultsPerPageInitialState,
@@ -208,3 +217,9 @@ export {
   SearchParameterManager,
   buildSearchParameterManager,
 } from './search-parameter-manager/headless-search-parameter-manager';
+
+export {
+  SearchStatus,
+  SearchStatusState,
+  buildSearchStatus,
+} from './search-status/headless-search-status';