+++ packages/samples/headless-react/src/App.tsx
@@ -1,5 +1,9 @@
 import logo from './logo.svg';
 import './App.css';
+import {useEffect} from 'react';
+
+import {Tab} from './components/tab/tab.class';
+import {Tab as TabFn} from './components/tab/tab.fn';
 import {SearchBox} from './components/search-box/search-box.class';
 import {SearchBox as SearchBoxFn} from './components/search-box/search-box.fn';
 import {DidYouMean} from './components/did-you-mean/did-you-mean.class';
@@ -12,7 +16,16 @@ import {ResultList} from './components/result-list/result-list.class';
 import {ResultList as ResultListFn} from './components/result-list/result-list.fn';
 import {Pager} from './components/pager/pager.class';
 import {Pager as PagerFn} from './components/pager/pager.fn';
+import {ResultsPerPage} from './components/results-per-page/results-per-page.class';
+import {ResultsPerPage as ResultsPerPageFn} from './components/results-per-page/results-per-page.fn';
+import {engine} from './engine';
+import {Section} from './layout/section';
+import {QuerySummary} from './components/query-summary/query-summary.class';
+import {QuerySummary as QuerySummaryFn} from './components/query-summary/query-summary.fn';
+import {Facet} from './components/facet/facet.class';
+import {Facet as FacetFn} from './components/facet/facet.fn';
 import {
+  buildTab,
   buildSearchBox,
   buildDidYouMean,
   buildQueryError,
@@ -28,14 +41,23 @@ import {
   buildResultsPerPage,
   buildPager,
 } from '@coveo/headless';
-import {ResultsPerPage} from './components/results-per-page/results-per-page.class';
-import {ResultsPerPage as ResultsPerPageFn} from './components/results-per-page/results-per-page.fn';
-import {engine} from './engine';
-import {Section} from './layout/section';
-import {QuerySummary} from './components/query-summary/query-summary.class';
-import {QuerySummary as QuerySummaryFn} from './components/query-summary/query-summary.fn';
-import {Facet} from './components/facet/facet.class';
-import {Facet as FacetFn} from './components/facet/facet.fn';
+import {bindSearchParametersToURI} from './components/search-parameter-manager/search-parameter-manager';
+
+const tabs = {
+  all: buildTab(engine, {
+    initialState: {isActive: true},
+    options: {expression: ''},
+  }),
+  messages: buildTab(engine, {
+    options: {expression: '@objecttype==Message'},
+  }),
+  confluence: buildTab(engine, {
+    options: {
+      expression:
+        '@connectortype==Confluence2Crawler AND NOT @documenttype==Space',
+    },
+  }),
+};
 
 const searchBox = buildSearchBox(engine, {options: {numberOfSuggestions: 8}});
 
@@ -68,11 +90,29 @@ const resultsPerPage = buildResultsPerPage(engine, {
 
 const pager = buildPager(engine, {options: {numberOfPages: 6}});
 
+const {autoUpdateURI: startUpdatingURI} = bindSearchParametersToURI(engine);
+
 function App() {
+  useEffect(() => startUpdatingURI(), []);
+
   return (
     <div className="App">
       <header className="App-header">
         <img src={logo} className="App-logo" alt="logo" />
+        <Section title="tabs">
+          <nav>
+            <Tab active>All</Tab>
+            <Tab expression="@objecttype==Message">Messages</Tab>
+            <Tab expression="@connectortype==Confluence2Crawler AND NOT @documenttype==Space">
+              Confluence
+            </Tab>
+          </nav>
+          <nav>
+            <TabFn controller={tabs.all}>All</TabFn>
+            <TabFn controller={tabs.messages}>Messages</TabFn>
+            <TabFn controller={tabs.confluence}>Confluence</TabFn>
+          </nav>
+        </Section>
         <Section title="search-box">
           <SearchBox />
           <SearchBoxFn controller={searchBox} />