+++ packages/samples/headless-react/src/components/query-summary/query-summary.fn.tsx
@@ -16,19 +16,31 @@ export const QuerySummary: FunctionComponent<QuerySummaryProps> = (props) => {
   useEffect(() => controller.subscribe(() => setState(controller.state)), []);
 
   const {
+    hasResults,
     hasQuery,
+    hasDuration,
     firstResult,
     lastResult,
     total,
     query,
     durationInSeconds,
   } = state;
-  return hasQuery ? (
-    <p>
-      Results {firstResult}-{lastResult} of {total} for {query} in{' '}
-      {durationInSeconds} seconds
-    </p>
-  ) : null;
+
+  if (!hasResults) {
+    return null;
+  }
+
+  const summary = [`Results ${firstResult}-${lastResult} of ${total}`];
+
+  if (hasQuery) {
+    summary.push(`for ${query}`);
+  }
+
+  if (hasDuration) {
+    summary.push(`in ${durationInSeconds} seconds`);
+  }
+
+  return <p>{summary.join(' ')}</p>;
 };
 
 // usage